MAINT: Move diagnostics to main
Move diagnostics out of sandbox
